[sw,lib] Fix debug log argument errors in tests
Fix all compilation errors in bazel targets that occur as a result
of the compile-time logging argument checking that ensures that
uint64_t values are not passed to `LOG` or equivalent functions
(any functions like `CHECK` or `LOG_INFO` that call LOG).

This can be verified by running

```
./bazelisk.sh build -t- --keep_going --define bitstream=skip
	//util/... //sw/...
```

and then grepping the output for the specific error (for 'must be
no wider than 32 bits').

All fixes simply involve performing an appropriate cast from a 64
bit integer type to a relevant 32 bit integer type. It is assumed
based on the context of these variables that all values are safe
to cast without any loss of information; no existing formatting
utilities exist to print integers larger than 32 bits.

+++ sw/device/lib/testing/aon_timer_testutils.c
@@ -23,7 +23,7 @@ status_t aon_timer_testutils_get_aon_cycles_32_from_us(uint64_t microseconds,
                                  /*rem_out=*/NULL);
   TRY_CHECK(cycles_ <= UINT32_MAX,
             "The value 0x%08x%08x can't fit into the 32 bits timer counter.",
-            (cycles_ >> 32), (uint32_t)cycles_);
+            (uint32_t)(cycles_ >> 32), (uint32_t)cycles_);
   *cycles = (uint32_t)cycles_;
   return OK_STATUS();
 }
@@ -41,7 +41,7 @@ status_t aon_timer_testutils_get_us_from_aon_cycles(uint64_t cycles,
                              /*rem_out=*/NULL);
   TRY_CHECK(uss <= UINT32_MAX,
             "The value 0x%08x%08x can't fit into the 32 bits timer counter.",
-            (uss >> 32), (uint32_t)uss);
+            (uint32_t)(uss >> 32), (uint32_t)uss);
   *us = (uint32_t)uss;
   return OK_STATUS();
 }
